+++ h3ron-graph/src/algorithm/mod.rs
@@ -4,3 +4,9 @@ mod dijkstra;
 pub mod path;
 pub mod shortest_path;
 pub mod within_weight_threshold;
+
+// re-export all algorithm traits
+pub use covered_area::CoveredArea;
+pub use differential_shortest_path::DifferentialShortestPath;
+pub use shortest_path::{ShortestPath, ShortestPathManyToMany};
+pub use within_weight_threshold::WithinWeightThreshold;